fix(security): move entity-taxonomy helpers out of 'use server' module (#427)
Five helpers previously exported from actions/taxonomy.ts had no auth
check:
  - getEntityTaxonomyDefinitions(organizationId, entityType)
  - getAllEntityTaxonomyDefinitions(organizationId)
  - getEntityTaxonomyValues(organizationId, entityType, entityId)
  - syncEntityTaxonomyValues(organizationId, entityType, entityId, termIds)
  - getEntityTaxonomyValuesForMany(organizationId, entityType, entityIds)

Each accepted a caller-supplied organizationId, and because they lived in
a 'use server' module, every authenticated user could call them directly
with any orgId — reading or modifying entity-taxonomy state for any
tenant.

Fix: relocate the helpers to lib/entity-taxonomy-helpers.ts. Files in
lib/ are not 'use server' modules, so the helpers cannot be reached as
RPC endpoints by construction. Each call site is a server action that
has already authenticated the actor and verified ownership of the
relevant entity.

Same architecture as #412 (cross-org-link helpers in lib/).

Call sites updated:
  - 7 action files (adrs, applications, capabilities, initiatives,
    objectives, principles, services) — import path swapped from
    @/actions/taxonomy → @/lib/entity-taxonomy-helpers
  - 8 page files (capabilities, principles, taxonomy, objectives,
    initiatives, applications, adrs, services) — three of which
    needed the import line split because they imported a mix of
    helpers (now in lib) and other taxonomy reads (still in actions)
  - 1 API route (applications/export) — same path swap

Test added (tests/integration/entity-taxonomy-helpers.test.ts):
  - helpers are exported from lib/entity-taxonomy-helpers
  - actions/taxonomy no longer exports them — locks in the
    architecture; if anyone re-exports them, the test fails

All 213 integration tests pass; tsc --noEmit clean.

Capability: cms/iam, cms/multi-org
Closes #427
