+++ apps/govea/src/app/(admin)/services/page.tsx
@@ -2,7 +2,7 @@ import { auth } from '@/lib/auth'
 import { redirect } from 'next/navigation'
 import { getServices } from '@/actions/services'
 import { getPersonas } from '@/actions/personas'
-import { getEntityTaxonomyDefinitions, getEntityTaxonomyValuesForMany } from '@/actions/taxonomy'
+import { getEntityTaxonomyDefinitions, getEntityTaxonomyValuesForMany } from '@/lib/entity-taxonomy-helpers'
 import { ServiceTable } from './service-table'
 
 export default async function ServicesPage() {
